FIXED CHATBOT AND WEIGHT FORECASTING SYSTEM

+++ lib/widgets/main_screen_widgets/food_page_screen/meal_container.dart
@@ -4,10 +4,10 @@ import 'package:flutter/material.dart';
 class MealContainer extends StatelessWidget {
   final String foodName;
   final String foodDescription;
-  final int calories;
-  final int protein;
-  final int carbs;
-  final int fat;
+  final double calories;
+  final double protein;
+  final double carbs;
+  final double fat;
   final String serving;
   final void Function()? onPressed;
